lyzr_backend/requirements.txt
amqp==5.3.1
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
azure-core==1.35.0
azure-storage-blob==12.26.0
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
channels_redis==4.3.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
debugpy==1.8.15
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
django-redis==6.0.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.94.2
executing==2.2.0
feedparser==6.0.11
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kombu==5.5.4
matplotlib-inline==0.1.7
msgpack==1.1.1
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pyzmq==27.0.0
razorpay==1.4.2
redis==6.2.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
zope.interface==7.2
